feat: fetch per-variant data asynchronously
