+++ web/src/components/ClusterDistribution/ClusterDistributionPlotCard.tsx
@@ -4,7 +4,7 @@ import React from 'react'
 import { Card, CardBody, CardHeader, Col, Row } from 'reactstrap'
 import styled from 'styled-components'
 
-import type { ClusterDistributionDatum } from 'src/io/usePerClusterData'
+import type { ClusterDistributionDatum } from 'src/io/getPerClusterData'
 import { Link } from 'src/components/Link/Link'
 import { PlotCardTitle } from 'src/components/Common/PlotCardTitle'
 import { ClusterDistributionPlot } from 'src/components/ClusterDistribution/ClusterDistributionPlot'
