+++ web/src/state/ClustersForPerCountryData.ts
@@ -1,11 +1,11 @@
 import copy from 'fast-copy'
+import { updateUrlOnClustersSet } from 'src/state/Clusters'
 import type { Cluster } from 'src/state/Clusters'
-import { updateUrlQuery } from 'src/helpers/urlQuery'
 import { fetchPerCountryDataRaw } from 'src/io/getPerCountryData'
 import { atomFamilyAsync } from 'src/state/utils/atomAsync'
 
-export const clustersForPerCountryAtom = atomFamilyAsync<Cluster[], string>({
-  key: 'clustersForPerCountryAtom',
+export const clustersForPerCountryDataAtom = atomFamilyAsync<Cluster[], string>({
+  key: 'clustersForPerCountryDataAtom',
   async default(region: string) {
     const { regions } = await fetchPerCountryDataRaw()
     const perCountryData = regions.find((candidate) => candidate.region === region)
@@ -15,19 +15,5 @@ export const clustersForPerCountryAtom = atomFamilyAsync<Cluster[], string>({
     const clusterNames = copy(perCountryData.cluster_names).sort()
     return clusterNames.map((cluster) => ({ cluster, enabled: true }))
   },
-  effects: [
-    ({ onSet }) => {
-      onSet(onClustersSet)
-    },
-  ],
+  effects: [updateUrlOnClustersSet],
 })
-
-function onClustersSet(clusters: Cluster[]) {
-  // If all clusters are enabled, we will remove cluster url params
-  const hasAllEnabled = clusters.every((cluster) => cluster.enabled)
-
-  // eslint-disable-next-line no-void
-  void updateUrlQuery({
-    variant: hasAllEnabled ? [] : clusters.filter((cluster) => cluster.enabled).map((cluster) => cluster.cluster),
-  })
-}
