+++ web/src/state/utils/atomEffect.ts
@@ -1,0 +1,3 @@
+import type { AtomEffect } from 'recoil'
+
+export type AtomEffectParams<T> = Parameters<AtomEffect<T>>[0]
